add logic new context
