+++ api/mockData.ts
@@ -1,0 +1,32 @@
+export const taskMock = [
+  {
+    id: Date.now().toString(),
+    title: "Buy groceries",
+    completed: false,
+    createdAt: new Date(),
+  },
+  {
+    id: (Date.now() + 1).toString(),
+    title: "Finish React Native project",
+    completed: true,
+    createdAt: new Date(),
+  },
+  {
+    id: (Date.now() + 2).toString(),
+    title: "Call mom",
+    completed: false,
+    createdAt: new Date(),
+  },
+  {
+    id: (Date.now() + 3).toString(),
+    title: "Read a book",
+    completed: true,
+    createdAt: new Date(),
+  },
+  {
+    id: (Date.now() + 4).toString(),
+    title: "Clean the house",
+    completed: false,
+    createdAt: new Date(),
+  },
+];
