--- screens/Home/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useTaskManager } from "./useTaskManager";
